requirements.txt
amqp==2.2.2
billiard==3.5.0.3
celery==4.1.0
certifi==2018.1.18
chardet==3.0.4
Django==2.0
Fabric3==1.13.1.post1
gunicorn==19.7.1
idna==2.6
kombu==4.1.0
#netifaces==0.10.6
Pillow==5.0.0
psycopg2==2.7.3.1
PyJWT==1.6.0
pytz==2018.3
requests==2.18.4
urllib3==1.22
vine==1.1.4
